refactor get license payload to another util
